x86, gart: mark reserve memory block as nomap
There have been a few workarounds for the GART aperture, including
56dd669a138c ("Insert GART region into resource map"),
2a3e83c6f96c ("x86/gart: Exclude GART aperture from vmcore"), all of
them is preventing kernel from touching the aperture memory. But there
is still memory remaining. Currently if a userspace program tries to
use /proc/kcore to read the content of kernel memory space, kernel may
crash if there are memory reserved for GART iommu.

Samiliar workaround like 2a3e83c6f96c could be done, but instead, we
could mark that area as MEMORYBLOCK_NOMAP instead, as /proc/kcore will
skip unmapped memory area, the problem is fixed.

We are using memblock_reserve to reverse memory for the aperature, this
commit marks the reserved memory block as nomap, make sure the kernel is
unable to touch that area, and void further problems.

This fix the problem that kernel panic when userspace program try to