hyperv_fb: let hyperv_fb work with 64bit framebuffer address
This is a fix for kexec reboot. Kexec will generate boot_params for next
booting up kernel and will pass existing framebuffer address for second
kernel direct. This make it possible that a 64bit framebuffer address is
used even if the hypervisor won't give a 64bit buffer for ordinary
bootup.

Without this fix, hyperv_fb will condiser a truncated address as the
framebuffer address used by efi_fb and try to release it which will lead
to unexpected behavior. Thing will not just work with this patch but it
make the failback went smooth and avoid warning or hanging issue.