kexec/kdump: allow userspace to view and reduce low crash memory
When the crash memory is reserved above 4G, a dedicated low memory region
need to be reserved for second kernel, or else second kernel will fail
due to lack of low memory.

We already have a /sys/kernel/kexec_crash_size available for user to
view and reduce reserved memory, but there is no way to do the same thing
for reserved low memory. So implement it for low memory too.