Reserve extra region when swiotlb is enabled
To ensure second kernel will always work, an extra region for swiotlb is
required. This region have to meet 4 conditions:

1. Below 4G.
2. Just enough for SWIOTLB, so it won't waste low memory.
3. Except below 4G, it also have to be lower than normal crashkernel
   memory, else it will break all existing version of kexec-tools.
4. Don't alloc extra memory for SWIOTLB for crash kernel, unless needed.

So try to alloc a swiotlb region (which is most likely smaller than the
normal crashkernel region) below the pre-reserved crashkernel region
when swiotlb is being initialized.

Also vastly reuse previous low memory reserve logic. So if swiotlb is
initialized early and not being freed it can also be reserved early. And
when crashkernel=xxx,high is used the behavior will be exactly the same
with before.

This only take care of ordinary swiotlb init and skips late swiotlb init
case, since it uses page allocator asking for DMA pages, and it's rare.

Signed-off-by: Kairui Song <kasong@redhat.com>