Reserve extra region when swiotlb is enabled
To ensure second kernel will always work, try to alloc an extra region
for swiotlb when swiotlb in enabled. This helps to avoid OOM issues when
user is not aware of the availibility of swiotlb.

This region have to meet 4 conditions:
1. Below 4G.
2. Just enough for SWIOTLB, so it won't waste low memory.
3. Except below 4G, it also have to be lower than normal crashkernel
   memory, else it will break all existing version of kexec-tools.
4. Don't alloc extra memory for SWIOTLB for crash kernel until it's
   really needed.

So try to alloc a swiotlb region (which is most likely smaller than the
normal crashkernel region) below the pre-reserved crashkernel region
when swiotlb is being initialized.

And alloc a seperate region will always work better than alloc a whole
region with larger size. If there isn't a big enough hole in memory
allocating a large chunk of memory may fail, and as long as allocating a
large size of memory can work, alloc a seperate region will also
garenteen to work.

And crashkernel reserve happens very early by which time the kernel
can't judge if swiotlb will be in used. So if the automatical allocation
of swiotlb region for crash kernel will be done in kernel space, it have
to be done after iommu allocation is done, by which time the the kernel
will be aware of if swiotlb might be used.

And kernel also need to check and free it in late initcalls, as some
iommu will get initialized very late and release the swiotlb. In such
case the second kernel will most likely do not need such extra region.
As even if swiotlb get initialized in early stage in crash kernel, it
will be release before the kernel run into high memory pressure.

Also reuse the low memory resource. As when crashkernl=xxx,high is in
use the low memory resource is always ensured to be enough including
swiotlb usage.

And only take care of ordinary swiotlb init and skips late swiotlb init
case, since it uses page allocator asking for DMA pages, and it's rare.

Signed-off-by: Kairui Song <kasong@redhat.com>