pci: disable master bit in late init state for kdump
Kdump problem about DMA has been there for a long time. That is,
when a kernel is switched to the kdump kernel, DMA derived from first
kernel is not stopped. In most cases, this is not harmful as kdump
kernel runs in pre-reserved memory region.

But the ongoing DMA could ruin the old memory and corrupt the vmcore,
and these devices are out of control of the kdump kernel, which is risky
after all.

Ideally, devices should be stopped as early as possible, like at the
beginning of the kdump kernel boot. But that is very arch dependent and
less doable. This patch try to stop the DMAs in a later stage in PCI
subsystem.

is used for PCI passthrough on KVM guest. In the case of the machine I
use, when intel_iommu=on is specified, DMAR error is detected in kdump
kernel and PCI SERR is also detected. Finally kdump fails because some
devices does not work correctly.

The root cause is that ongoing DMA from first kernel causes DMAR fault
because page table of DMAR is initialized while kdump kernel is booting
up. Therefore to address this problem DMA needs to be stopped before
DMAR is initialized at kdump kernel boot time. By these patches, PCIe
devices are reset by hot reset and its DMA is stopped when reset_devices
is specified. One problem of this solution is that the monitor blacks
out when VGA controller is reset. So this patch does not reset the port
whose child endpoint is VGA device.

What I tried:

Signed-off-by: Kairui Song <kasong@redhat.com>