watchdog/hpwdt: Set kdumptimeout to 0 when already in kdump
If the it's already in kdump kernel, the kdumptimeout parameter is
meanless. And if kdumptimeout is set to -1, watchdog is stopped upon a
NMI from the firmware, preventing it from resetting the kernel in case
of a kernel hanging issue.

I've produced such senarior, kdump kernel encounter an hardware
issue and hanged. An hardware error NMI is triggered by the firmware,
the hpwdt NMI handler in hpwdt caught that before the hang, and stopped
the watchdog as the kdumptimeout is -1. Now the system is completely
dead, the watchdog is stopped so nothing will reset it.

Just let it be 0 in kdump kernel to prevent such issue.

Signed-off-by: Kairui Song <kasong@redhat.com>