efi: fix memmap splitting function
Currently when splitting a efi memmap entry, only the physical memory
address are re-calculated. The virt address of the splitted entries
are zero-ed.

Right now, it works because kernel will alloc the virt address upon calling
efi_enter_virtual_mode. So as long as the memmap split happens before
entering virtual mode, it works.

But kexec'ed kernel could fail, because after kexec, efi is already in
virtual mode so kernel simply do fixed mapping reuse the recorded virt
address. If the virt address is zero-ed, the mapping will be invalid.

Update the virt_addaress as well when splitting a memmap entry to fix
this potential issue.

Signed-off-by: Kairui Song <kasong@redhat.com>