mm, swap: use an enum to define all cluster flags and wrap flags changes
Currently we are only using flags to indicate which list the cluster is
on, using one bit for each list type might be a waste as the list type
grows we will consume too much bits. And current the mixed usage of "&"
and "==" is a bit confusing.

Make it clean by using an enum to define all possible cluster status,
only an off-list cluster will have the NONE (0) flag. And use
a wrapper to annotate and sanitize all flag setting and list movement.

Suggested-by: Chris Li <chrisl@kernel.org>
Signed-off-by: Kairui Song <kasong@tencent.com>