workingset: simplify and use a more intuitive model
This commit basically removed workingset_activation and reduced calls
to workingset_age_nonresident.

The idea behind this change is a new way to calculate the refault
distance and prepare for adapting refault distance for workingset
tracking.

Currently, refault distance based reactivation for help keep workingset
pages in memory fpr active/inactive LRU, it works by estimating the
re-access distance of a page, if it's small enough, then reactivate it
on fault.

As described in mm/workingset.c, this is based on two assumptions:
1. Activation of an inactive page will left-shift LRU pages (considering
   LRU starts from right).
2. Eviction of an inactive page will left-shift LRU pages.

Assumption 2 is correct, but assumption 1 is not always true, an activated
page could be anywhere in the LRU list (through mark_page_accessed), it
only left-shift the pages on its right side. Besides, one page can get
activate/deactivated for many times.

So instead we can have a simpler idea here: Just presume the evicted
pages are still in memory, each has an corresponding eviction timestamp
(nonresistence_age) that is increased and recorded upon each eviction,
just like before. These timestamp could logically form a "Shadow LRU"
(not really a LRU but hotness of the page is well represented by the
timestamp), Let the `nonresistence_age` still be NA, then we have:

  Let SP = ((NA's reading @ now) - (NA's reading @ eviction time))

                           +----- memory available ----+
                           |                           |
 +-------------------------+===============+===========+
 | *   shadows  O O  O     |   INACTIVE    |   ACTIVE  |
 +-+-----------------------+===============+===========+
   |                       |
   +-----------------------+
   |         SP
re-fault page     O -> Logical hole left by re-faulted pages.
                       On a real LRU entries should be moved upon
                       access but this is actually a read-only
                       and imaginary queue of shadows, so consider
                       these faulted pages left a place holder hold.
                  * -> The page corresponding to SP

The reactivation rule is similar to before, with slight adjustments.

It can be easily seen that SP stands for the offset of a page in the
"Shadow LRU", which is also how far the current workflow could push
a page out of available memory in unit of pages. Since all evicted
page was once head of INACTIVE list, the estimated minimum value of
re-access distance is:

  SP + NR_INACTIVE

On re-fault, the page *may* get activated and stay in memory if we put
it to active LRU if:

  SP + NR_INACTIVE < NR_INACTIVE + NR_ACTIVE

Which can be simplified to:

  SP < NR_ACTIVE

And because there are two types of pages, we want to sacrifice the
inactive part of opposite type so cold pages are reclaimed more
proactively, so it will be:

  SP(anon/file) < NR_ACTIVE(anon + file) + NR_INACTIVE(file/anon)

If it evaluated to true, then the page is worth getting reactivated.

Since this is only an estimation, based on several hypotheses, and
it could break the ability of LRU to distinguish workingset pages out
of caches. In extreme cases all refault causing activation will lead to
worse thrashing, so throttle this by two factors naturally:

1. Notice previously re-faulted in pages may leave "holes" on the "shadow
   LRU", that part is left unhandled on purpose to decrease reactivate
   rate for pages that have a large SP value (the larger SP value a page
   has, the more likely it will be affected by such holes).
2. Shadow will be reclaimed, by list LRU and shrinker, so when under
   heavy pressure long evicted pages will never be reactivated.

Code-wise, this is simpler than before since no longer need to do memcg
lruvec workingset data update when activating a page, and so far,
benchmarks are showing a similar or better result. And the performance
is also better even if there is no memory pressure since some memcg
iteration and atomic operation is no longer needed for ordinary routine.

Some test results:

  memtier test (with 16G ramdisk as swap and 4G memcg limit on an i7-9700):
  memcached -u nobody -m 16384 -s /tmp/memcached.socket \
    -a 0766 -t 12 -B binary &
  memtier_benchmark -S /tmp/memcached.socket -P memcache_binary -n allkeys\
    --key-minimum=1 --key-maximum=32000000 --key-pattern=P:P -c 1 \
    -t 12 --ratio 1:0 --pipeline 8 -d 2000 -x 6

  fio test 1 (with 16G ramdisk on 28G VM on an i7-9700):
  fio -name=refault --numjobs=12 --directory=/mnt --size=1024m \
    --buffered=1 --ioengine=io_uring --iodepth=128 \
    --iodepth_batch_submit=32 --iodepth_batch_complete=32 \
    --rw=randread --random_distribution=random --norandommap \
    --time_based --ramp_time=5m --runtime=5m --group_reporting

  fio test 2 (with 16G ramdisk on 28G VM on an i7-9700):
  fio -name=mglru --numjobs=10 --directory=/mnt --size=1536m \
    --buffered=1 --ioengine=io_uring --iodepth=128 \
    --iodepth_batch_submit=32 --iodepth_batch_complete=32 \
    --rw=randread --random_distribution=zipf:1.2 --norandommap \
    --time_based --ramp_time=10m --runtime=5m --group_reporting

  mysql (using oltp_read_only from sysbench, with 12G of buffer pool
  in a 10G memcg):
  sysbench /usr/share/sysbench/oltp_read_only.lua <auth and db params> \
    --tables=36 --table-size=2000000 --threads=12 --time=1800

  kernel build test done with 3G memcg limit on an i7-9700.

Before (Average of 6 test run):
fio: IOPS=5125.5k
fio2: IOPS=7291.16k
memcached: 57600.926 ops/s
mysql: 6491.5 tps
kernel-build: 1817.13499 seconds

After (Average of 6 test run):
fio: IOPS=5137.5k (+2.3%)
fio2: IOPS=7300.67k (+1.3%)
memcached: 57878.422 ops/s (+4.8%)
mysql: 6491.1 tps (~0.0%)
kernel-build: 1813.66231 seconds (+2.0%)

Signed-off-by: Kairui Song <kasong@tencent.com>