mm, swap: remove old allocation path for HDD
We are currently using different swap allocation algorithm for HDD and
non-HDD. This leads to the existing of different set of locking, and the
code path is heavily bloated, causing troubles for further optimization
and maintenance.

This commit removes all HDD swap allocation and related dead code, and
use cluster allocation algorithm instead.

The performance may drop a little bit temporarily, and should be
negligible: The main advantage of legacy HDD allocation algorithm
is that is tend to use continuous slots, but swap device gets fragmented
quickly anyway, and the attempt to use continuous slots will fail easily.

This commit also enables mTHP swap on HDD, which should be beneficial,
and following commits will adapt and optimize the cluster allocator
for HDD.

Suggested-by: Chris Li <chrisl@kernel.org>
Suggested-by: "Huang, Ying" <ying.huang@intel.com>
Signed-off-by: Kairui Song <kasong@tencent.com>