mm, swap: clean up plist removal and adding
When swap device is full (inuse_pages == pages), it should be
removed from the plist. And if any slot is freed, the swap device should
be added back to the plist. On swapoff / swapon, the swap device will
also be force removed / added.

This is currently serialized by si->lock, and some historical sanity
check code are still here. This commit decouple this from the protection
of si->lock and clean it up to prepare for si->lock rework.

Noticing inuse_pages counter is the only thing decides if a device
should be removed from or added to the plist (except swapon / swapoff
as a special case), and inuse_pages is a very hot counter. So to avoid
extra overhead on the counter update hot path, and make it possible to
check and update the plist when the counter value changes, embed the
plist state into the inuse_pages counter, and turn the counter into
an atomic. This way we can check and update the counter with one CAS
and avoid any extra synchronization.

If the counter is full (inuse_pages == pages) with the off-list bit
unset, try to remove it from the plist. If the counter is not full
(inuse_pages != pages) with the off-list bit set, try to add it to
the plist. Removing and adding is serialized with lock as well as the
bit setting. Ordinary counter updates will be lockless.

Signed-off-by: Kairui Song <kasong@tencent.com>