mm: folio_contains shouldn't be called for swap cache folios
Currently, none of the folio_contains user is suppose to be used for
swap cache.

Swap cache mapping is kind of special, they don't share the same logic
with page cache, eg. the index with the swap cache space is not linearly
mapped to the swapfile offset, and many filemap operations doesn't apply
for swap cache mapping.

So make it explicit that folio_contains should not be used for swap
cache folios.

Signed-off-by: Kairui Song <kasong@tencent.com>