mm, swap: rename folio_free_swap to folio_reclaim_swap
The function name is confusing, as it doesn't free the swap space if
there is any data still in the swap device, it only reclaim the swap
cache space if the folio is not swapped out. So rename it to a proper
name and make place for a real folio_free_swap later.

Signed-off-by: Kairui Song <kasong@tencent.com>