mm, swap: use cluster table for swap cache with and switch API
At this point it looks like just a downgrade from Xarray to flat
array, but based on this, later commits will implement a cluster
based unified swap map, with dynamic allocation, which should reduce the
memory usage while make the performance better.

And this downgrade shows a performance gain, and get rid of many
address_spaces instances of one swap device.

Which is about >40% faster in system time in multiple tests.

Signed-off-by: Kairui Song <kasong@tencent.com>