mm, swap: add generic check and sanity check for swap exclusive
To prepare for removing the SWP_SYNCHRONOUS_IO check, introduce a swap
exclusive check with sanity checks and check.

For a folio being swapped in, it should either be mapped entirely or
only one of it sub page will be mapped.

swap_pte_batch already ensures a mTHP being swapped in

Signed-off-by: Kairui Song <kasong@tencent.com>