SEEMS OK: mm/workingset: leave highest 8 bits empty for anon shadow
This should be OK for foreseeable future, take 52 bits physical address
space as example, for 4K pages, there would be at most 40 bits of
addressable pages. Currently we have 36 bits available (with NODES_SHIFT
set to 10, this can be decreased for more bits), so in worst case
refault distance compare will be done for every 64K sized bucket. This
increases the bucket size to 16M, which actually should be fine as the
workingset size will be way larger than the bucket size for such large
machines.

For MGLRU 28 bits can track a huge amount of gens already, there should
be no problem either.

Signed-off-by: Kairui Song <kasong@tencent.com>