mm/swap: rename __read_swap_cache_async to swap_cache_alloc_folio
__read_swap_cache_async is widely used to allocate and ensure a folio is
in swapcache, or get the folio if a folio is already there.

It's not async, and it's not doing any read. Rename it to better
present its usage, and prepare to be reworked as part of new swap
cache APIs.

Also add some comments for the function. The skip_if_exists argument is
a workaround that will be dropped very soon so just mention it too.

Signed-off-by: Kairui Song <kasong@tencent.com>