mm, swap: sanitize swap cache lookup convention
Swap cache lookup is lock less, the returned folio could be invalidated
any time before locking it. So the caller always have to lock and check
the folio before use.

Introduce a helper for swap cache folio verification, and document this
convention, avoid touching the folio until the folio has been verified.

And update all current users using this convention, and use the new
added helper when possible for easier debugging.

Some current users like swapoff path is OK with an invalid folio,
because during swapoff, other folios can't be re-swapped out again
using the same device / entry, so a invalid folio should always mean the
PTE is gone and so the PTE check will fail before using this folio.
But it's kind of fragile, and checking folio before locking the PTE might
be helpful for reducing overhead, and this makes the convention easier
to follow.

Signed-off-by: Kairui Song <kasong@tencent.com>